+++ src/configuration/SensorConfig.cpp
@@ -41,6 +41,8 @@ const char* calibrationConfigTypeToString(SensorConfigType type) {
 			return "SoftFusion (common)";
 		case SensorConfigType::BNO0XX:
 			return "BNO0XX";
+		case SensorConfigType::RUNTIME_CALIBRATION:
+			return "SoftFusion (runtime calibration)";
 		default:
 			return "UNKNOWN";
 	}
