+++ src/debug.h
@@ -100,8 +100,8 @@
 #define FIRMWARE_VERSION "UNKNOWN"
 #endif
 
-#ifndef USE_NONBLOCKING_CALIBRATION
-#define USE_NONBLOCKING_CALIBRATION true
+#ifndef USE_RUNTIME_CALIBRATION
+#define USE_RUNTIME_CALIBRATION true
 #endif
 
 #define DEBUG_MEASURE_SENSOR_TIME_TAKEN true
